+++ src/dawmans/index/lexical.py
@@ -1,0 +1,1 @@
+"""The bm25s wrapper."""